feat(solved.ac): 창문닫기

+++ java/src/약수배수와소수2/다음소수/Main.java
@@ -12,7 +12,7 @@ class Main {
 		}
 		while (true) {
 			boolean flag = true;
-			for (int i = 2; i <= (int) Math.sqrt(result); i++) {
+			for (long i = 2; i <= (long) Math.sqrt(result); i++) {
 				if (result % i == 0) {
 					result++;
 					flag = false;
